# backport-2.1: sql: point spans should never include interleaves