--- a/pkg/sql/logictest/testdata/logic_test/interleaved
+++ b/pkg/sql/logictest/testdata/logic_test/interleaved
@@ -577,14 +577,17 @@ r1: sending batch 1 Scan to (n1,s1):1
 fetched: /a/primary/1 -> NULL
 Del /Table/79/1/1/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/1/#/80/1{-/#}
 querying next range at /Table/79/1/1/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/1/#/80/1/1{-/#}
 querying next range at /Table/79/1/1/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/1/#/80/1/1/0
 querying next range at /Table/79/1/1/#/80/1/1/0
 r1: sending batch 1 Del, 1 BeginTxn to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/1/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/1/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/1/#/80/1/1/#/81/1
@@ -594,14 +597,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/2 -> NULL
 Del /Table/79/1/2/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/2/#/80/1{-/#}
 querying next range at /Table/79/1/2/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/2/#/80/1/1{-/#}
 querying next range at /Table/79/1/2/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/2/#/80/1/1/0
 querying next range at /Table/79/1/2/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/2/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/2/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/2/#/80/1/1/#/81/1
@@ -611,14 +617,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/3 -> NULL
 Del /Table/79/1/3/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/3/#/80/1{-/#}
 querying next range at /Table/79/1/3/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/3/#/80/1/1{-/#}
 querying next range at /Table/79/1/3/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/3/#/80/1/1/0
 querying next range at /Table/79/1/3/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/3/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/3/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/3/#/80/1/1/#/81/1
@@ -628,14 +637,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/4 -> NULL
 Del /Table/79/1/4/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/4/#/80/1{-/#}
 querying next range at /Table/79/1/4/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/4/#/80/1/1{-/#}
 querying next range at /Table/79/1/4/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/4/#/80/1/1/0
 querying next range at /Table/79/1/4/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/4/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/4/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/4/#/80/1/1/#/81/1
@@ -645,14 +657,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/5 -> NULL
 Del /Table/79/1/5/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/5/#/80/1{-/#}
 querying next range at /Table/79/1/5/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/5/#/80/1/1{-/#}
 querying next range at /Table/79/1/5/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/5/#/80/1/1/0
 querying next range at /Table/79/1/5/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/5/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/5/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/5/#/80/1/1/#/81/1
@@ -662,14 +677,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/6 -> NULL
 Del /Table/79/1/6/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/6/#/80/1{-/#}
 querying next range at /Table/79/1/6/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/6/#/80/1/1{-/#}
 querying next range at /Table/79/1/6/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/6/#/80/1/1/0
 querying next range at /Table/79/1/6/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/6/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/6/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/6/#/80/1/1/#/81/1
@@ -679,14 +697,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/7 -> NULL
 Del /Table/79/1/7/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/7/#/80/1{-/#}
 querying next range at /Table/79/1/7/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/7/#/80/1/1{-/#}
 querying next range at /Table/79/1/7/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/7/#/80/1/1/0
 querying next range at /Table/79/1/7/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/7/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/7/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/7/#/80/1/1/#/81/1
@@ -696,14 +717,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/8 -> NULL
 Del /Table/79/1/8/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/8/#/80/1{-/#}
 querying next range at /Table/79/1/8/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/8/#/80/1/1{-/#}
 querying next range at /Table/79/1/8/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/8/#/80/1/1/0
 querying next range at /Table/79/1/8/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/8/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/8/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/8/#/80/1/1/#/81/1
@@ -713,14 +737,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/9 -> NULL
 Del /Table/79/1/9/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/9/#/80/1{-/#}
 querying next range at /Table/79/1/9/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/9/#/80/1/1{-/#}
 querying next range at /Table/79/1/9/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/9/#/80/1/1/0
 querying next range at /Table/79/1/9/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/9/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/9/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/9/#/80/1/1/#/81/1
@@ -730,14 +757,17 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/10 -> NULL
 Del /Table/79/1/10/0
 cascading delete into table: 80 using index: 1
+CascadeScan /Table/79/1/10/#/80/1{-/#}
 querying next range at /Table/79/1/10/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
+CascadeScan /Table/79/1/10/#/80/1/1{-/#}
 querying next range at /Table/79/1/10/#/80/1/1
 r1: sending batch 1 Scan to (n1,s1):1
 Del /Table/79/1/10/#/80/1/1/0
 querying next range at /Table/79/1/10/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
+CascadeScan /Table/79/1/10/#/80/1/1/#/81/1{-/#}
 querying next range at /Table/79/1/10/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/10/#/80/1/1/#/81/1