--- a/pkg/sql/logictest/testdata/logic_test/interleaved
+++ b/pkg/sql/logictest/testdata/logic_test/interleaved
@@ -577,7 +577,7 @@ r1: sending batch 1 Scan to (n1,s1):1
 fetched: /a/primary/1 -> NULL
 Del /Table/79/1/1/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/1/#/80/1{-/#}
+CascadeScan /Table/79/1/1/#/80/{1-2}
 querying next range at /Table/79/1/1/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/1/#/80/1/1{-/#}
@@ -587,7 +587,7 @@ Del /Table/79/1/1/#/80/1/1/0
 querying next range at /Table/79/1/1/#/80/1/1/0
 r1: sending batch 1 Del, 1 BeginTxn to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/1/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/1/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/1/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/1/#/80/1/1/#/81/1
@@ -597,7 +597,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/2 -> NULL
 Del /Table/79/1/2/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/2/#/80/1{-/#}
+CascadeScan /Table/79/1/2/#/80/{1-2}
 querying next range at /Table/79/1/2/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/2/#/80/1/1{-/#}
@@ -607,7 +607,7 @@ Del /Table/79/1/2/#/80/1/1/0
 querying next range at /Table/79/1/2/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/2/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/2/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/2/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/2/#/80/1/1/#/81/1
@@ -617,7 +617,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/3 -> NULL
 Del /Table/79/1/3/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/3/#/80/1{-/#}
+CascadeScan /Table/79/1/3/#/80/{1-2}
 querying next range at /Table/79/1/3/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/3/#/80/1/1{-/#}
@@ -627,7 +627,7 @@ Del /Table/79/1/3/#/80/1/1/0
 querying next range at /Table/79/1/3/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/3/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/3/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/3/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/3/#/80/1/1/#/81/1
@@ -637,7 +637,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/4 -> NULL
 Del /Table/79/1/4/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/4/#/80/1{-/#}
+CascadeScan /Table/79/1/4/#/80/{1-2}
 querying next range at /Table/79/1/4/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/4/#/80/1/1{-/#}
@@ -647,7 +647,7 @@ Del /Table/79/1/4/#/80/1/1/0
 querying next range at /Table/79/1/4/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/4/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/4/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/4/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/4/#/80/1/1/#/81/1
@@ -657,7 +657,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/5 -> NULL
 Del /Table/79/1/5/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/5/#/80/1{-/#}
+CascadeScan /Table/79/1/5/#/80/{1-2}
 querying next range at /Table/79/1/5/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/5/#/80/1/1{-/#}
@@ -667,7 +667,7 @@ Del /Table/79/1/5/#/80/1/1/0
 querying next range at /Table/79/1/5/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/5/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/5/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/5/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/5/#/80/1/1/#/81/1
@@ -677,7 +677,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/6 -> NULL
 Del /Table/79/1/6/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/6/#/80/1{-/#}
+CascadeScan /Table/79/1/6/#/80/{1-2}
 querying next range at /Table/79/1/6/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/6/#/80/1/1{-/#}
@@ -687,7 +687,7 @@ Del /Table/79/1/6/#/80/1/1/0
 querying next range at /Table/79/1/6/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/6/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/6/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/6/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/6/#/80/1/1/#/81/1
@@ -697,7 +697,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/7 -> NULL
 Del /Table/79/1/7/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/7/#/80/1{-/#}
+CascadeScan /Table/79/1/7/#/80/{1-2}
 querying next range at /Table/79/1/7/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/7/#/80/1/1{-/#}
@@ -707,7 +707,7 @@ Del /Table/79/1/7/#/80/1/1/0
 querying next range at /Table/79/1/7/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/7/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/7/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/7/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/7/#/80/1/1/#/81/1
@@ -717,7 +717,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/8 -> NULL
 Del /Table/79/1/8/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/8/#/80/1{-/#}
+CascadeScan /Table/79/1/8/#/80/{1-2}
 querying next range at /Table/79/1/8/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/8/#/80/1/1{-/#}
@@ -727,7 +727,7 @@ Del /Table/79/1/8/#/80/1/1/0
 querying next range at /Table/79/1/8/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/8/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/8/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/8/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/8/#/80/1/1/#/81/1
@@ -737,7 +737,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/9 -> NULL
 Del /Table/79/1/9/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/9/#/80/1{-/#}
+CascadeScan /Table/79/1/9/#/80/{1-2}
 querying next range at /Table/79/1/9/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/9/#/80/1/1{-/#}
@@ -747,7 +747,7 @@ Del /Table/79/1/9/#/80/1/1/0
 querying next range at /Table/79/1/9/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/9/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/9/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/9/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/9/#/80/1/1/#/81/1
@@ -757,7 +757,7 @@ r1: sending batch 1 Scan, 1 QueryIntent to (n1,s1):1
 fetched: /a/primary/10 -> NULL
 Del /Table/79/1/10/0
 cascading delete into table: 80 using index: 1
-CascadeScan /Table/79/1/10/#/80/1{-/#}
+CascadeScan /Table/79/1/10/#/80/{1-2}
 querying next range at /Table/79/1/10/#/80/1
 r1: sending batch 1 Scan to (n1,s1):1
 CascadeScan /Table/79/1/10/#/80/1/1{-/#}
@@ -767,7 +767,7 @@ Del /Table/79/1/10/#/80/1/1/0
 querying next range at /Table/79/1/10/#/80/1/1/0
 r1: sending batch 1 Del to (n1,s1):1
 cascading delete into table: 81 using index: 1
-CascadeScan /Table/79/1/10/#/80/1/1/#/81/1{-/#}
+CascadeScan /Table/79/1/10/#/80/1/1/#/81/{1-2}
 querying next range at /Table/79/1/10/#/80/1/1/#/81/1
 r1: sending batch 1 Scan to (n1,s1):1
 querying next range at /Table/79/1/10/#/80/1/1/#/81/1