… pkg/bench/bench_test.go
… pkg/sql/sqlbase/index_encoding.go
… pkg/sql/tablewriter_upsert.go